backend/pyproject.toml
[tool.poetry]
name = "worker-daemon-backend"
version = "0.1.0"
description = ""
authors = ["Zygimantas Straznickas <hi@zygi.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.3.4"
pydantic = "^2.7.4"
fastapi = "^0.111.0"
watchfiles = "^0.22.0"
sse-starlette = "^2.1.2"
# tender = {path = "../../codegen"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
